+++ client/src/pages/Comment.js
@@ -1,132 +1,42 @@
-var CommentBox = React.createClass({
-    loadCommentsFromServer: function() {
-       $.ajax({
-         url: this.props.url,
-         dataType: 'json',
-         cache: false,
-         success: function(data) {
-           this.setState({data: data});
-         }.bind(this),
-         error: function(xhr, status, err) {
-           console.error(this.props.url, status, err.toString());
-         }.bind(this)
-       });
-     },
-     getInitialState: function() {
-       return {data: []};
-     },
-      componentDidMount: function() {
-       this.loadCommentsFromServer();
-       setInterval(this.loadCommentsFromServer,   this.props.pollInterval);
-     },
-     render: function() {
-       return (
-         <div className="commentBox">
-           <h1>Comments</h1>
-           <CommentList data={this.state.data} />
-           <CommentForm />
-         </div>
-       );
-     }
-   });
-   ReactDOM.render(
-     <CommentBox url="/api/comments" pollInterval={2000} />,
-     document.getElementById('content')
-   );
+class MarkdownEditor extends React.Component {
+  constructor(props) {
+    super(props);
+    this.md = new Remarkable();
+    this.handleChange = this.handleChange.bind(this);
+    this.state = { value: 'Hello, **world**!' };
+  }
 
+  handleChange(e) {
+    this.setState({ value: e.target.value });
+  }
 
+  getRawMarkup() {
+    return { __html: this.md.render(this.state.value) };
+  }
 
+  render() {
+    return (
+      <div className="MarkdownEditor">
+        <h3>Comment</h3>
+        <label htmlFor="markdown-content">
+          Enter comments
+        </label>
+        <textarea
+          id="markdown-content"
+          onChange={this.handleChange}
+          defaultValue={this.state.value}
+        />
+        <h3></h3>
+        <div
+          className="content"
+          dangerouslySetInnerHTML={this.getRawMarkup()}
+        />
+      </div>
+    );
+  }
+}
 
-var CommentForm = React.createClass({
-    getInitialState: function() {
-      return {author: '', text: ''};
-    },
-    handleAuthorChange: function(e) {
-      this.setState({author: e.target.value});
-    },
-    handleTextChange: function(e) {
-      this.setState({text: e.target.value});
-    },
-    handleSubmit: function(e) {
-      e.preventDefault();
-      var author = this.state.author.trim();
-      var text = this.state.text.trim();
-      if (!text || !author) {
-        return;
-      }
-      this.props.onCommentSubmit({author: author, text: text});
-      this.setState({author: '', text: ''});
-    },
-    render: function() {
-      return (
-        <form className="commentForm" onSubmit={this.handleSubmit}>
-          <input
-            type="text"
-            placeholder="Your name"
-            value={this.state.author}
-            onChange={this.handleAuthorChange}
-          />
-          <input
-            type="text"
-            placeholder="Say something..."
-            value={this.state.text}
-            onChange={this.handleTextChange}
-          />
-          <input type="submit" value="Post" />
-        </form>
-      );
-    }
-  });
-  var CommentBox = React.createClass({
-    loadCommentsFromServer: function() {
-      $.ajax({
-        url: this.props.url,
-        dataType: 'json',
-        cache: false,
-        success: function(data) {
-          this.setState({data: data});
-        }.bind(this),
-        error: function(xhr, status, err) {
-          console.error(this.props.url, status, err.toString());
-        }.bind(this)
-      });
-    },
-    handleCommentSubmit: function(comment) {
-      var comments = this.state.data;
-      // Optimistically set an id on the new comment. It will be replaced by an
-      // id generated by the server. In a production application you would likely
-      // not use Date.now() for this and would have a more robust system in place.
-      comment.id = Date.now();
-      var newComments = comments.concat([comment]);
-      this.setState({data: newComments});
-      $.ajax({
-        url: this.props.url,
-        dataType: 'json',
-        type: 'POST',
-        data: comment,
-        success: function(data) {
-          this.setState({data: data});
-        }.bind(this),
-        error: function(xhr, status, err) {
-          this.setState({data: comments});
-          console.error(this.props.url, status, err.toString());
-        }.bind(this)
-      });
-    },
-    getInitialState: function() {
-      return {data: []};
-    },
-    componentDidMount: function() {
-      this.loadCommentsFromServer();
-      setInterval(this.loadCommentsFromServer, this.props.pollInterval);
-    },
-    render: function() {
-      return (
-        <div className="commentBox">
-          <h1>Comments</h1>
-          <CommentList data={this.state.data} />
-          <CommentForm onCommentSubmit={this.handleCommentSubmit} />
-        </div>
-      );
-    }
-  });
+ReactDOM.render(
+  <MarkdownEditor />,
+  document.getElementById('markdown-example')
+);
